wip: rework translation from frontend to cir
This commit introduces a Core Syntax Tree (CST) layer between AST
and CIR, meant to represent the syntax tree after semantic
analysis, on which pattern match compilation, and other
simplifications and rewrites are performed, analogous to Core Erlang.

CST has both an "internal" representation, prefixed with `I` that
represents CST nodes which have not yet been fully transformed and thus
have more complex syntax than the final simplified CST which is much
more strict.

+++ compiler/binary/src/spec.rs
@@ -6,14 +6,14 @@ pub enum BinaryEntrySpecifier {
     Integer {
         signed: bool,
         endianness: Endianness,
-        unit: i64,
+        unit: usize,
     },
     Float {
         endianness: Endianness,
-        unit: i64,
+        unit: usize,
     },
     Binary {
-        unit: i64,
+        unit: usize,
     },
     Utf8,
     Utf16 {
@@ -37,7 +37,7 @@ impl BinaryEntrySpecifier {
         }
     }
 
-    pub fn unit(&self) -> i64 {
+    pub fn unit(&self) -> usize {
         match self {
             Self::Integer { unit, .. } | Self::Float { unit, .. } | Self::Binary { unit, .. } => {
                 *unit
